[v1.0.7] - separated `<Sidebar />` to independent component - added dropdown menu for `menu`   - list all menu items   - add new menu - added storybook for sidebar   - same as `<PageWrapper />`
